The config writer moves to the crate that owns the schema
It began as part of `robotctl configure`, when that was the only thing writing
`robotd.toml`. Two more callers have appeared since — `robotctl policy` and
`robotctl pad` write keys of their own — and a daemon serving `pad.bind` over
the radio needs the same writer, because `padd` re-reads `[pad]` every second
and a binding that only changed something in memory would be reverted before
anybody let go of the phone.

So it is `robotd_params::edit` now: `Row`, `Edit`, `Model`, and the pad helpers.
The guarantee worth keeping is that nothing writes a file `robotd` would refuse
to start on, and that guarantee is only as good as its least careful writer — so
the writer belongs beside the schema it validates against rather than in one of
its callers. `toml_edit` is pure Rust, so this crate's rule that nothing in it
grows a C toolchain or a network stack still holds.

The twenty tests that exercise the model moved with it, which was the point of
doing this properly rather than re-exporting and leaving them: `robotd-params`
is the crate on the recovery path, and shipping a writer there with its tests in
another crate would be the wrong way round. `from_text` is public now so a
caller can build a model over text it has already read — mostly used for testing,
which is a good enough reason on its own: a config written as a string says what
a case is about better than a temporary file does.

What stayed in `robotctl` is what is genuinely an operator tool's: which systemd
unit a change needs restarted, restarting it, the divergence listing and the
full-screen editor. `configure.rs` goes from 1680 lines to 794.

One test came back after going the wrong way. `a_save_remembers_what_it_wrote_so
_the_right_daemon_restarts` reads as a model test and is a restart-mapping test;
it needed `units_to_restart`, which is exactly the line between the two modules.
